requirements.txt
pydantic==1.10.2
slist==0.1.2
python-dotenv==0.21.0
pytest==7.2.0
retry==0.9.2
mypy == 0.991
types-retry==0.9.9
types-requests==2.28.11.5
requests==2.28.1
pandas==1.5.2
openai==0.23.1
neptune-client==0.16.9